Move cold-brew-test.js into src directory, update package.json to point at the new location, update test file require statement to point at new location
