+++ package.json
@@ -2,7 +2,7 @@
   "name": "cold-brew",
   "version": "1.0.35",
   "description": "A testing module for TDD with apps that use the RTCPeerConnection API",
-  "main": "cold-brew-test.js",
+  "main": "src/cold-brew-test.js",
   "scripts": {
     "test": "mocha test/*.spec.js",
     "generate-docs": "./node_modules/.bin/jsdoc ./cold-brew-test.js -d ./docs",
